requirements.txt
yeelight
zwift-client
